refactoring for better naming and javadoc added
